convert2sql/requirements.txt
aspose-cells
aspose-cells-python
tabula-py
csv2sql